feat(frontend): Phase 5 - Migrate dashboard/agents_behavior/sync to V2 structure
Migrated 3 pages to V2 directory structure:
- dashboard.js (465→5 files): register + page + StatsSection + ActivityFeed + TrendChart
- agents_behavior.js (661→6 files): register + page + constants + PersonalityTab + BehaviorLog + StatsTab
- sync.js (576→8 files): register + page + SyncStatus + AutoSync + ManualControls + BackendConfig + HotUpdate + Changelog

Key improvements:
- ErrorHandler.wrap() on all register.js
- Lazy loading via dynamic import()
- data-action event delegation replacing inline onclick
- Dashboard: 15s polling + SSE real-time updates preserved
- Agents Behavior: SSE event dispatch + hover delegation
- Sync: module cross-communication via modules object

Updated index.html: replaced 3 old script refs with register.js refs
