+++ frontend/index.html
@@ -187,13 +187,13 @@
 <script src="/js/utils/confirm-dialog.js" defer></script>
 <script src="/js/api.js" defer></script>
 <script src="/js/components.js" defer></script>
-<script src="/js/pages/dashboard.js" defer></script>
+<script src="/js/pages/dashboard/register.js" defer></script>
 <script src="/js/pages/knowledge/register.js" defer></script>
 <script src="/js/pages/sessions/register.js" defer></script>
 <script src="/js/pages/chat/register.js" defer></script>
 <script src="/js/pages/memory/register.js" defer></script>
 <script src="/js/pages/cron/register.js" defer></script>
-<script src="/js/pages/agents_behavior.js" defer></script>
+<script src="/js/pages/agents_behavior/register.js" defer></script>
 <script src="/js/pages/config/register.js" defer></script>
 <!-- V2 Page Modules (Phase 1) -->
 <script src="/js/pages/about/register.js" defer></script>
@@ -203,7 +203,7 @@
 <!-- V2 Page Modules End -->
 <script src="/js/pages/marketplace/register.js" defer></script>
 <script src="/js/pages/logs/register.js" defer></script>
-<script src="/js/pages/sync.js" defer></script>
+<script src="/js/pages/sync/register.js" defer></script>
 <script src="/js/app.js" defer></script>
 
 <!-- 全局错误边界：JS 报错时显示友好提示而不是白屏 -->
